Add Sentinel support
Sentinels are used for health checking, fallback strategy and master
IP address discovery. Latter is necessary in a Master/Slave/Sentinel
setup to retrieve the data of the correct Redis node (master) for the
exporter.

Two command line parameters are introduced: 'sentinel.addr' and
'sentinel.mastr'. Both are string slices and need to have the same
number of strings, since for master discovery we need the address of one
sentinel and the name of the master to discover.

If the parameters are provided and the master discovery was successful,
the IP address(es) of the discovered master(s) are appended to the
addresses given by 'redis.addr'.

If 'redis.addr' is not specified only the discovered addresses are taken
into account.

Closes #59