Refactor to allow viewing of multiples redises simultaneously.
To view a redis cluster as a whole, the drop down is converted to a multi value. Panels where there are lots of metrics calclate them as a sum across all redis instances. When readable, value per redis instance is used. Single stat panels use max for memory and uptime and sum across all instances for clients connected.