Add configuration to set TLS minVersion on server
This adds the configuration `tls-server-min-version` to be able
to set the minimum TLS version that will be accepted by the metrics
server port.
Since TLS 1.0 and 1.1 have reached end-of-life they are normally not allowed
within some organisations, but Go enables it by default on TLS servers.
With this new configuration only TLS1.2/TLS1.3 clients are accepted by default,
but can be opened up to accept the less secure version via config.